Move state-churning tools to own module
