+++ tests/integration/churn_state/Nymph.sol
@@ -1,0 +1,9 @@
+pragma solidity ^0.5.7;
+
+// Nymph exists purely to be deleted later
+
+contract Nymph {
+    function poof() public {
+        selfdestruct(msg.sender);
+    }
+}
